Separated out helper methods into partial class
