+++ CSharp/FizzBuzz/Program.cs
@@ -2,56 +2,34 @@
 
 namespace FizzBuzz
 {
-    class Program
+    partial class Program
     {
-        static void Main(string[] args)
+        /// <summary>
+        /// The Main method is the starting point of your program and is the first code to be run
+        /// </summary>
+        static void Main()
         {
             Console.WriteLine("Welcome to FizzBuzz!");
 
-            int x = InputNumber("Please enter a start number and then press the 'Enter' key: ");
-            int y = InputNumber("Please enter an end number and then press the 'Enter' key: ");
-            
+            int x = 1;
+            int y = 20;
+
+            // we now 'call' the PrintNumberRange method, passing in x as the startNumber parameter
+            // and 'y' as the endNumber parameter.
             PrintNumberRange(x, y);
 
             Console.WriteLine("Program has completed. Press any key to exit");
             Console.ReadKey();
         }
 
+        /// <summary>
+        /// This is the method in which you will complete your exercise
+        /// </summary>
+        /// <param name="startNumber">This is that start number you will pass in</param>
+        /// <param name="endNumber">This is the end number you will pass in</param>
         static void PrintNumberRange(int startNumber, int endNumber)
         {
             // START here!            
         }
-
-#region Helper Methods - DO NOT make any changes inside here, else you could break your program!
-        static int InputNumber(string text)
-        {
-            Console.WriteLine(text);
-            var args = Console.ReadLine();
-
-            if (!int.TryParse(args, out var number))
-            {
-                Console.WriteLine($"Oops! {args} doesn't look like a number. Try again!");
-                InputNumber(text);                
-            }
-
-            return number;
-        }
-
-        static void PrintNumber(int number)
-        {
-            Console.Write(number.ToString());
-        }
-
-        static void PrintWord(string word)
-        {
-            Console.Write(word);
-        }
-
-        static void PrintSpace()
-        {
-            Console.Write(" ");
-        }
     }
-#endregion 
-    
 }
